% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemocratPlatforms.R
\docType{data}
\name{DemocratPlatforms}
\alias{DemocratPlatforms}
\title{Democrat Party Platforms}
\format{
A data frame
\describe{
\item{X}{Line ID}
\item{head}{Heading of the Section}
\item{Year}{Year of the Election}
\item{Party}{Democrat}
\item{Date}{Date of the platform, as listed on UCSB site}
\item{Candidate}{Nominee for the election}
\item{topic}{General Topic of the section -- Economy, Society, Foreign Affair, Government, Introduction}
\item{text}{Actual text minus headings}
}
}
\source{
\url{https://www.presidency.ucsb.edu/documents/presidential-documents-archive-guidebook/national-political-party-platforms}
}
\usage{
DemocratPlatforms
}
\description{
Party Platforms for the Democrat party from 2000 to 2016
}
\examples{
\dontrun{DemocratPlatforms}

}
\keyword{datasets}
