% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC08.R
\docType{data}
\name{dnc08}
\alias{dnc08}
\title{Democrat National Convention 2008}
\format{
a Large Corpus with 17 texts
}
\source{
Varies. See data metafile for details.
}
\usage{
dnc08
}
\description{
Captures the speeches given at the 2008 Democrat
National Convention
}
\examples{
\dontrun{dnc08}


}
\keyword{datasets}
