% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC08.R
\docType{data}
\name{dnc08_meta}
\alias{dnc08_meta}
\title{Metadata File for the 2008 DNC}
\format{
a data frame with 17 rows and 8 variables
\describe{
\item{Last}{Last Name of Speaker}
\item{First}{First Name of Speaker}
\item{Filename}{Corresponding File name to \code{\link{dnc08}}}
\item{Day}{Day in the convention the speech was given}
\item{Type}{speech or benediction}
\item{Source}{Provider of the document}
\item{Convention}{Type of convention -- DNC}
\item{URL}{Link to source}
}
}
\usage{
dnc08_meta
}
\description{
Metadata File for the 2008 DNC
}
\examples{
dnc08_meta
}
\keyword{datasets}
