% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC12.R
\docType{data}
\name{dnc12}
\alias{dnc12}
\title{Democrat National Convention 2012}
\format{
a Large Corpus with 111 texts
}
\source{
Varies. See data metafile for details.
}
\usage{
dnc12
}
\description{
Captures the speeches given at the 2012 Democrat
National Convention
}
\examples{
\dontrun{dnc12}


}
\keyword{datasets}
