% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC16.R
\docType{data}
\name{dnc16}
\alias{dnc16}
\title{Democrat National Convention 2016}
\format{
a Large Corpus with 155 texts
}
\source{
Varies. \url{https://www.politico.com/story/2016/07/dnc-2016-schedule-of-events-and-speakers-225617}
}
\usage{
dnc16
}
\description{
Captures the speeches given at the 2016 Democrat
National Convention
}
\examples{
\dontrun{dnc16}


}
\keyword{datasets}
