% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC16.R
\docType{data}
\name{dnc16_meta}
\alias{dnc16_meta}
\title{Metadata File for the 2016 DNC}
\format{
a data frame with 155 rows and 7 variables
\describe{
\item{Last}{Last Name of Speaker}
\item{First}{First Name of Speaker}
\item{Filename}{Corresponding File name to \code{\link{dnc16}}}
\item{Day}{Day in the convention the speech was given}
\item{Type}{speech or benediction}
\item{Source}{Provider of the document}
\item{Convention}{Type of convention -- DNC}
}
}
\usage{
dnc16_meta
}
\description{
Metadata File for the 2016 DNC
}
\examples{
dnc16_meta
}
\keyword{datasets}
