% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNC20.R
\docType{data}
\name{dnc20}
\alias{dnc20}
\title{Democrat National Convention 2020}
\format{
a Data Frame with 1693 observations
\describe{
\item{Speaker}{Last Name of Speaker}
\item{Day}{Day in the convention the speech was given}
\item{Date}{Date of the Speech}
\item{Convention}{Type of convention -- DNC}
\item{text}{Text of speech}
}
}
\source{
Varies. Nightly Transcripts from Rev
}
\usage{
dnc20
}
\description{
Captures the speeches given at the 2020 Democrat
National Convention
}
\details{
Due to the online nature of the convention, the speeches
included in this data set are those included in the 9-11 PM
EST prime time livestream

Links to the original speech transcripts from Rev are as follows:

Night 1: \url{https://www.rev.com/blog/transcripts/democratic-national-convention-dnc-night-1-transcript}

Night 2: \url{https://www.rev.com/blog/transcripts/democratic-national-convention-dnc-2020-night-2-transcript}

Night 3: \url{https://www.rev.com/blog/transcripts/democratic-national-convention-dnc-night-3-transcript}

Night 4: \url{https://www.rev.com/blog/transcripts/2020-democratic-national-convention-dnc-night-4-transcript}
}
\examples{
\dontrun{dnc20}


}
\keyword{datasets}
