% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\name{ExportingData}
\alias{ExportingData}
\alias{exportDoc}
\title{Export Functions for General Texts}
\usage{
exportDoc(df, dest, ...)
}
\arguments{
\item{df}{Data frame to export}

\item{dest}{Location and file name to save the data}
}
\description{
Export to plain text functions for plain text with settings
pre-configured globally
}
