% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC08.R
\docType{data}
\name{rnc08}
\alias{rnc08}
\title{Republican National Convention 2008}
\format{
a Large Corpus with 10 texts
}
\source{
Varies. See data metafile for details.
}
\usage{
rnc08
}
\description{
Captures the speeches given at the 2008 Republican
National Convention
}
\examples{
\dontrun{rnc08}

}
\keyword{datasets}
