% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC08.R
\docType{data}
\name{rnc08_meta}
\alias{rnc08_meta}
\title{Metadata File for the 2008 RNC}
\format{
a data frame with 10 rows and 8 variables
\describe{
\item{Last}{Last Name of Speaker}
\item{First}{First Name of Speaker}
\item{Filename}{Corresponding File name to \code{\link{rnc08}}}
\item{Day}{Day in the convention the speech was given}
\item{Type}{speech or benediction}
\item{Source}{Provider of the document}
\item{Convention}{Type of convention -- RNC}
\item{URL}{Link to source}
}
}
\usage{
rnc08_meta
}
\description{
Metadata File for the 2008 RNC
}
\examples{
rnc08_meta
}
\keyword{datasets}
