% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC12.R
\docType{data}
\name{rnc12}
\alias{rnc12}
\title{Republican National Convention 2012}
\format{
a Large Corpus with 33 texts
}
\source{
Varies. See data metafile for details.
}
\usage{
rnc12
}
\description{
Captures the speeches given at the 2012 Republican
National Convention
}
\examples{
\dontrun{rnc12}

}
\keyword{datasets}
