% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC12.R
\docType{data}
\name{rnc12_meta}
\alias{rnc12_meta}
\title{Metadata File for the 2012 RNC}
\format{
a data frame with 33 rows and 8 variables
\describe{
\item{Last}{Last Name of Speaker}
\item{First}{First Name of Speaker}
\item{Filename}{Corresponding File name to \code{\link{rnc12}}}
\item{Day}{Day in the convention the speech was given}
\item{Type}{speech or benediction}
\item{Source}{Provider of the document}
\item{Convention}{Type of convention -- RNC}
\item{URL}{Link to source}
}
}
\usage{
rnc12_meta
}
\description{
Metadata File for the 2012 RNC
}
\examples{
rnc12_meta
}
\keyword{datasets}
