% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC16.R
\docType{data}
\name{rnc16}
\alias{rnc16}
\title{Republican National Convention 2016}
\format{
a Large Corpus with 65 texts
}
\source{
Varies. \url{https://www.politico.com/story/2016/07/rnc-2016-schedule-of-events-and-speakers-225704}
}
\usage{
rnc16
}
\description{
Captures the speeches given at the 2016 Republican
National Convention
}
\examples{
\dontrun{rnc16}

}
\keyword{datasets}
