% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC16.R
\docType{data}
\name{rnc16_meta}
\alias{rnc16_meta}
\title{Metadata File for the 2016 RNC}
\format{
a data frame with 65 rows and 7 variables
\describe{
\item{Last}{Last Name of Speaker}
\item{First}{First Name of Speaker}
\item{Filename}{Corresponding File name to \code{\link{rnc16}}}
\item{Day}{Day in the convention the speech was given}
\item{Type}{speech or benediction}
\item{Source}{Provider of the document}
\item{Convention}{Type of convention -- RNC}
}
}
\usage{
rnc16_meta
}
\description{
Metadata File for the 2016 RNC
}
\examples{
rnc16_meta
}
\keyword{datasets}
