% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapeDebate.R
\name{ScraperDebates}
\alias{ScraperDebates}
\alias{scrapeDebUCSB}
\alias{scrapeDebOrg}
\title{Scraper Functions for Debates}
\usage{
scrapeDebUCSB(url)

scrapeDebOrg(url)
}
\arguments{
\item{url}{A URL in quotes}
}
\value{
text string in the format
\code{SPEAKERA: text SPEAKERB: text} with
line breaks between the speakers
}
\description{
Functions to scrape Debates from
\code{https://www.presidency.ucsb.edu} or
\code{https://www.debates.org/}
}
\details{
Use \code{scrapeDebUCSB} for texts from \code{https://www.presidency.ucsb.edu}
and use \code{scrapeDebOrg} for texts from
\code{https://www.debates.org/voter-education/debate-transcripts/}. Since
the websites are set up a bit differently, the functions are
adjusted to account for the difference CSS labels of the main
text. Both return a text string which can then be manipulated
as desired.
}
\examples{
# Debates from https://www.presidency.ucsb.edu
url <- "https://www.presidency.ucsb.edu/documents/presidential-debate-washington-university-st-louis-missouri"
Debate <- scrapeDebUCSB(url) # Scrape 2016 Presidential Debate

# Debates from Debates.org
url <- "https://www.debates.org/voter-education/debate-transcripts/september-26-2016-debate-transcript/"
Debate <- scrapeDebOrg(url)

}
