% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNC16.R
\name{rnc16}
\alias{rnc16}
\title{Republican National Convention 2016}
\arguments{
\item{rnc16_meta}{Metadata file for convention. Identifies speakers,
day, filename and source}
}
\description{
Captures the speeches given at the 2016 Republican
National Convention
}
